requirements.txt
streamlit
pillow
sentence-transformers
torch
scikit-learn